% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plogr-package.R
\docType{package}
\name{plogr-package}
\alias{plogr}
\alias{plogr-package}
\title{The 'plog' C++ Logging Library}
\description{
'plog' is a simple header-only logging library.
Add LinkingTo: plogr in your package to use it.
}
\examples{
plogr_demo <- Rcpp::cppFunction(depends = "plogr", '
// C++ code begin
#include <plogr.h>

RObject plogr_demo() {
  plog::init_r(plog::info);
  LOG_INFO << "shown";
  LOG_DEBUG << "not shown";
  plog::init_r("DEBUG");
  LOG_DEBUG << "shown now";
  return R_NilValue;
}

#include <Rcpp.h> // not necessary to use plogr
// C++ code end
'
)

plogr_demo()
}

